Controller-Service/requirements.txt
bottle
googleads
google-cloud-bigquery
google-cloud-firestore
google-cloud-logging
google-cloud-tasks
